Lesson 2: Shared constructor on models.
