+++ src/Model/Comment.php
@@ -2,27 +2,8 @@
 
 namespace Masterclass\Model;
 
-use PDO;
-
-class Comment
+class Comment extends BaseModel
 {
-    /**
-     * @var PDO
-     */
-    protected $db;
-
-    /**
-     * Comment constructor.
-     * @param array $config
-     */
-    public function __construct($config)
-    {
-        $dbconfig = $config['database'];
-        $dsn = 'mysql:host=' . $dbconfig['host'] . ';dbname=' . $dbconfig['name'];
-        $this->db = new PDO($dsn, $dbconfig['user'], $dbconfig['pass']);
-        $this->db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-    }
-
     /**
      * Create new comment.
      * @param string $username
