+++ src/Model/Story.php
@@ -4,25 +4,8 @@ namespace Masterclass\Model;
 
 use PDO;
 
-class Story
+class Story extends BaseModel
 {
-    /**
-     * @var PDO
-     */
-    protected $db;
-
-    /**
-     * Story constructor.
-     * @param array $config
-     */
-    public function __construct($config)
-    {
-        $dbconfig = $config['database'];
-        $dsn = 'mysql:host=' . $dbconfig['host'] . ';dbname=' . $dbconfig['name'];
-        $this->db = new PDO($dsn, $dbconfig['user'], $dbconfig['pass']);
-        $this->db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-    }
-
     /**
      * Create new story.
      * @param string $username
